Refactor/buildingrule/#153 팀빌딩 규칙 컨트롤러 리펙토링 및 리퀘스트매핑 수정

+++ team/src/main/java/sansam/team/buildingrule/command/infrastructure/repository/JpaBuildingRuleRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import sansam.team.buildingrule.command.domain.aggregate.BuildingRule;
 import sansam.team.buildingrule.command.domain.repositroy.BuildingRuleRepository;
 
-public interface JpaBuildingRuleRepository extends BuildingRuleRepository, JpaRepository<BuildingRule, Integer> {
+public interface JpaBuildingRuleRepository extends BuildingRuleRepository, JpaRepository<BuildingRule, Long> {
 }
